# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Binar Car Rental RESTful API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Muhlis Ahmad Abdillah",
      "email": "muhllis.ahmad@gmail.com"
    }
  },
  "servers": [
    {
      "description": "Binar Car Rental RESTful API",
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "carId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Car Id for return",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "carCategory": {
        "name": "category",
        "in": "path",
        "required": true,
        "description": "Car category",
        "schema": {
          "type": "string",
          "enum": ["small", "medium", "large"]
        }
      },
      "sortQuery": {
        "name": "sort",
        "in": "query",
        "required": false,
        "description": "Sort data according to a certain field",
        "schema": {
          "type": "string",
          "default": "-updated_at"
        },
        "examples": {
          "updated descending": {
            "description": "Sort data by the updated date descending",
            "value": "-updated_at"
          },
          "updated ascending": {
            "description": "Sort data by the updated date ascending",
            "value": "updated_at"
          }
        }
      },
      "pageQuery": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Go to specific page",
        "schema": {
          "type": "number",
          "default": 1
        }
      },
      "sizeQuery": {
        "name": "size",
        "in": "query",
        "required": false,
        "description": "Limit the amount of data being return",
        "schema": {
          "type": "number",
          "default": 10
        }
      },
      "searchQuery": {
        "name": "q",
        "in": "query",
        "required": false,
        "description": "Keyword to search data",
        "schema": {
          "type": "string"
        },
        "examples": {
          "car name keyword": {
            "description": "Keyword for search car data",
            "value": "Mercedes Benz"
          },
          "email keyword": {
            "description": "Keyword for search user email order",
            "value": "john.doe@domain.com"
          }
        }
      },
      "startDateQuery": {
        "name": "start_date",
        "in": "query",
        "required": true,
        "description": "Start date to rent a car",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "finishDateQuery": {
        "name": "finish_date",
        "in": "query",
        "required": true,
        "description": "Finish date to rent a car",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "driverServiceQuery": {
        "name": "driver_service",
        "in": "query",
        "required": true,
        "description": "Driver service",
        "schema": {
          "type": "boolean"
        }
      },
      "capacityQuery": {
        "name": "capacity",
        "in": "query",
        "required": false,
        "description": "Minimum car capacity",
        "schema": {
          "type": "number",
          "minimum": 1,
          "maximum": 10
        }
      }
    },
    "schemas": {
...
      "CarResponseAdmin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "created_by": {
            "type": "string",
            "example": "jonathanjoestar@gmail.com"
          },
          "updated_by": {
            "type": "string",
            "nullable": true
          },
          "deleted_by": {
            "type": "string",
            "nullable": true
          },
          "manufacture": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "Supra"
          },
          "transmission": {
            "type": "string",
            "example": "Manual"
          },
          "year": {
            "type": "number",
            "example": 2002
          },
          "driver_service": {
            "type": "boolean"
          },
          "rent_per_day": {
            "type": "number",
            "example": 250000
          },
          "image": {
            "type": "string",
            "example": "https://res.cloudinary.com/de3zdoakr/image/upload/v1717698129/binar-car-rental/upload/data/car/G%20299%20T.png",
            "nullable": true
          },
          "capacity": {
            "type": "number",
            "example": 4
          },
          "type": {
            "type": "string",
            "example": "sedan"
          },
          "options": {
            "type": "string",
            "nullable": true
          },
          "specs": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCarRequest": {
        "type": "object",
        "properties": {
          "manufacture": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "transmission": {
            "type": "string"
          },
          "plate": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "driver_service": {
            "type": "boolean"
          },
          "rent_per_day": {
            "type": "number",
            "example": 250000
          },
          "capacity": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "example": "sedan"
          },
          "category": {
            "type": "string",
            "enum": ["small", "medium", "large"]
          },
          "options": {
            "type": "string"
          },
          "specs": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "car": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "manufacture",
          "model",
          "transmission",
          "plate",
          "year",
          "driver_service",
          "rent_per_day",
          "capacity",
          "type",
          "category",
          "description"
        ]
      },
...
      "Paging": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "default": 1
          },
          "size": {
            "type": "number",
            "default": 10
          },
          "total_page": {
            "type": "number",
            "example": 6
          }
        }
      }
    }
  },
  "paths": {
...
    "/admin/cars": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Admin"],
        "summary": "Get Car List",
        "parameters": [
          {
            "$ref": "#/components/parameters/sortQuery"
          },
          {
            "$ref": "#/components/parameters/pageQuery"
          },
          {
            "$ref": "#/components/parameters/sizeQuery"
          },
          {
            "$ref": "#/components/parameters/searchQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Success\t: Get list cars data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CarResponseAdmin"
                      }
                    },
                    "paging": {
                      "$ref": "#/components/schemas/Paging"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "The current user do not have the authorization of accessing this route"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "Car data not found"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "You are not logged in!"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Admin"],
        "summary": "Create New Car Data",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateCarRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created\t: Car data created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "$ref": "#/components/schemas/CarResponseAdmin"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "The current user do not have the authorization of accessing this route"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict\t: Car data already exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "Car with specific plate number alrady exist"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/cars/{category}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Admin"],
        "summary": "Get Car By Category",
        "parameters": [
          {
            "$ref": "#/components/parameters/carCategory"
          },
          {
            "$ref": "#/components/parameters/sortQuery"
          },
          {
            "$ref": "#/components/parameters/pageQuery"
          },
          {
            "$ref": "#/components/parameters/sizeQuery"
          },
          {
            "$ref": "#/components/parameters/searchQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Success\t: Get list cars data by category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CarResponseAdmin"
                      }
                    },
                    "paging": {
                      "$ref": "#/components/schemas/Paging"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "The current user do not have the authorization of accessing this route"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "Car data not found"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "You are not logged in!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cars/search": {
      "get": {
        "tags": ["Cars"],
        "summary": "Search Car For Customer",
        "parameters": [
          {
            "$ref": "#/components/parameters/startDateQuery"
          },
          {
            "$ref": "#/components/parameters/finishDateQuery"
          },
          {
            "$ref": "#/components/parameters/driverServiceQuery"
          },
          {
            "$ref": "#/components/parameters/capacityQuery"
          },
          {
            "$ref": "#/components/parameters/sortQuery"
          },
          {
            "$ref": "#/components/parameters/pageQuery"
          },
          {
            "$ref": "#/components/parameters/sizeQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Success\t: Get list cars data by category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CarResponseCustomer"
                      }
                    },
                    "paging": {
                      "$ref": "#/components/schemas/Paging"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "Car data not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cars/{id}": {
      "get": {
        "tags": ["Cars"],
        "summary": "Get Car By Id (Customer)",
        "parameters": [
          {
            "$ref": "#/components/parameters/carId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success\t: Get car data by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CarResponseCustomer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "fail" },
                    "message": {
                      "type": "string",
                      "example": "Car with given ID cannot be found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}